Refactored models and admin
