+++ admin/__init__.py
@@ -1,0 +1,3 @@
+from .base import QuxModelAdmin
+from .plus import QuxPlusModelAdmin
+from .audit import AuditSummaryAdmin, AuditDetailsAdmin
