+++ admin/plus.py
@@ -1,0 +1,23 @@
+from django.contrib import admin
+from .base import QuxModelAdmin
+
+
+class QuxPlusModelAdmin(QuxModelAdmin):
+    list_filter = ("is_deleted",)
+
+    # noinspection PyProtectedMember
+    def get_queryset(self, request):
+        """
+        Returns a QuerySet of all model instances that can be edited by the
+        admin site. This is used by changelist_view.
+        """
+
+        # Default: qs = self.model._default_manager.get_query_set()
+        qs = self.model._default_manager.all_with_deleted()
+
+        # TODO: this should be handled by some parameter to the ChangeList.
+        # () because *None is bad
+        ordering = self.ordering or ()
+        if ordering:
+            qs = qs.order_by(*ordering)
+        return qs
