+++ models/__init__.py
@@ -1,4 +1,12 @@
-from .base import *
+from .base import (
+    CoreModel,
+    CoreManager,
+    QuxModel,
+)
+from .plus import (
+    CoreModelPlus,
+    CoreModelPlusManager,
+)
 from .audit import (
     CoreModelAuditSummary,
     CoreModelAuditDetails,
